upload/requirements.txt
django
djangorestframework
psycopg2
celery